backend/requirements.txt
fastapi
uvicorn
pillow

numpy

torch
torchvision
torchaudio

python-multipart
scikit-learn
scikit-image
httpx
python-dotenv
matplotlib
dotenv
aiofiles